backend/requirements.txt
fastapi
uvicorn 
websockets
diffusers
huggingface_hub
xformer
accelerate
jinja2
transformers
peft
pydantic-settings
aiortc